+++ src/main/java/org/opensearch/neuralsearch/processor/normalization/ScoreNormalizationUtil.java
@@ -82,4 +82,78 @@ class ScoreNormalizationUtil {
         }
         scores.set(subQueryIndex, normalizedScore);
     }
+
+    /**
+     * Validate parameters for this technique. Following is example of structured parameters that we will validate
+     * {
+     *     "technique": "arithmetic_mean",  // top-level parameter 1
+     *     "details": {                     // top-level parameter 2
+     *         "weights": [1, 2, 3],        // nested parameter 1
+     *         "color": "green".            // nested parameter 2
+     *     }
+     * }
+     * for this example client should pass:
+     *     top-level parameters: ["technique", "details"]
+     *     nested parameters: ["details" -> ["weights", "color"]]
+     * @param actualParameters map of actual parameters in form of name-value
+     * @param supportedParametersTopLevel collection of top-level parameters that we should validate against
+     * @param supportedParametersNested map of nested parameters that we should validate against, key is one of top level parameters and value is set of allowed nested params
+     */
+    public void validateParameters(
+        final Map<String, Object> actualParameters,
+        final Set<String> supportedParametersTopLevel,
+        final Map<String, Set<String>> supportedParametersNested
+    ) {
+        if (Objects.isNull(actualParameters) || actualParameters.isEmpty()) {
+            return;
+        }
+        boolean hasUnknownParameters = false;
+        for (Map.Entry<String, Object> entry : actualParameters.entrySet()) {
+            String paramName = entry.getKey();
+            Object paramValue = entry.getValue();
+
+            if (!supportedParametersTopLevel.contains(paramName)) {
+                hasUnknownParameters = true;
+                continue;
+            }
+            if (paramValue instanceof Map) {
+                Map<String, Object> nestedParams = (Map<String, Object>) paramValue;
+                validateNestedParameters(nestedParams, supportedParametersNested.get(paramName));
+            } else if (paramValue instanceof List) {
+                for (Object item : (List<?>) paramValue) {
+                    if (item instanceof Map) {
+                        validateNestedParameters((Map<String, Object>) item, supportedParametersNested.get(paramName));
+                    } else {
+                        hasUnknownParameters = true;
+                    }
+                }
+            } else {
+                if (supportedParametersNested.isEmpty()) {
+                    continue;
+                }
+                hasUnknownParameters = true;
+            }
+        }
+        if (hasUnknownParameters) {
+            throw new IllegalArgumentException("unrecognized parameters in normalization technique");
+        }
+    }
+
+    private void validateNestedParameters(Map<String, Object> parameters, Set<String> supportedNestedParams) {
+        if (Objects.isNull(parameters) || parameters.isEmpty()) {
+            return;
+        }
+        boolean hasUnknownParameters = false;
+        for (Map.Entry<String, Object> entry : parameters.entrySet()) {
+            String paramName = entry.getKey();
+
+            if (Objects.nonNull(supportedNestedParams) && !supportedNestedParams.contains(paramName)) {
+                hasUnknownParameters = true;
+                break;
+            }
+        }
+        if (hasUnknownParameters) {
+            throw new IllegalArgumentException("unrecognized parameters in normalization technique");
+        }
+    }
 }