+++ src/test/java/org/opensearch/neuralsearch/processor/normalization/MinMaxScoreNormalizationTechniqueTests.java
@@ -28,6 +28,7 @@ import org.opensearch.search.SearchShardTarget;
 import static org.opensearch.neuralsearch.processor.normalization.MinMaxScoreNormalizationTechnique.MIN_SCORE;
 import static org.opensearch.neuralsearch.query.HybridQueryBuilder.MAX_NUMBER_OF_SUB_QUERIES;
 import static org.opensearch.neuralsearch.util.TestUtils.DELTA_FOR_SCORE_ASSERTION;
+import static org.opensearch.neuralsearch.util.TestUtils.PARAM_NAME_LOWER_BOUNDS;
 
 /**
  * Abstracts normalization of scores based on min-max method
@@ -274,86 +275,124 @@ public class MinMaxScoreNormalizationTechniqueTests extends OpenSearchQueryTestC
         assertEquals(1.0f, topDocs3.scoreDocs[0].score, DELTA_FOR_SCORE_ASSERTION); // doc1 in third subquery
     }
 
-    public void testMode_fromString_validValues() {
-        assertEquals(MinMaxScoreNormalizationTechnique.Mode.APPLY, MinMaxScoreNormalizationTechnique.Mode.fromString("apply"));
-        assertEquals(MinMaxScoreNormalizationTechnique.Mode.CLIP, MinMaxScoreNormalizationTechnique.Mode.fromString("clip"));
-        assertEquals(MinMaxScoreNormalizationTechnique.Mode.IGNORE, MinMaxScoreNormalizationTechnique.Mode.fromString("ignore"));
+    public void testLowerBoundsModeFromString_whenValidValues_thenSuccessful() {
+        assertEquals(
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.APPLY,
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.fromString("apply")
+        );
+        assertEquals(
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.CLIP,
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.fromString("clip")
+        );
+        assertEquals(
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.IGNORE,
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.fromString("ignore")
+        );
         // Case insensitive check
-        assertEquals(MinMaxScoreNormalizationTechnique.Mode.APPLY, MinMaxScoreNormalizationTechnique.Mode.fromString("APPLY"));
+        assertEquals(
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.APPLY,
+            MinMaxScoreNormalizationTechnique.LowerBound.Mode.fromString("APPLY")
+        );
     }
 
     public void testMode_fromString_invalidValues() {
         IllegalArgumentException exception = expectThrows(
             IllegalArgumentException.class,
-            () -> MinMaxScoreNormalizationTechnique.Mode.fromString("invalid")
+            () -> MinMaxScoreNormalizationTechnique.LowerBound.Mode.fromString("invalid")
         );
         assertEquals("invalid mode: invalid, valid values are: apply, clip, ignore", exception.getMessage());
     }
 
-    public void testMode_fromString_nullOrEmpty() {
+    public void testLowerBoundsModeFromString_whenNullOrEmpty_thenFail() {
         IllegalArgumentException nullException = expectThrows(
             IllegalArgumentException.class,
-            () -> MinMaxScoreNormalizationTechnique.Mode.fromString(null)
+            () -> MinMaxScoreNormalizationTechnique.LowerBound.Mode.fromString(null)
         );
         assertEquals("mode value cannot be null or empty", nullException.getMessage());
-
-        IllegalArgumentException emptyException = expectThrows(
-            IllegalArgumentException.class,
-            () -> MinMaxScoreNormalizationTechnique.Mode.fromString("")
-        );
-        assertEquals("mode value cannot be null or empty", emptyException.getMessage());
     }
 
-    public void testMode_normalize_apply() {
+    public void testLowerBounds_whenModeIsApply_thenSuccessful() {
         float score = 0.5f;
-        float minScore = 0.2f;
+        float minScore = 0.1f;
         float maxScore = 0.8f;
         float lowerBoundScore = 0.3f;
 
-        float normalizedScore = MinMaxScoreNormalizationTechnique.Mode.APPLY.normalize(score, minScore, maxScore, lowerBoundScore);
+        float normalizedScore = MinMaxScoreNormalizationTechnique.LowerBound.Mode.APPLY.normalize(
+            score,
+            minScore,
+            maxScore,
+            lowerBoundScore
+        );
+        // we expect score as 0.5 - 0.3 / 0.8 - 0.3 = 0.2 / 0.5 = 0.4
         assertEquals(0.4f, normalizedScore, DELTA_FOR_SCORE_ASSERTION);
 
         // Test when score is below lower bound
-        float lowScore = 0.1f;
-        float normalizedLowScore = MinMaxScoreNormalizationTechnique.Mode.APPLY.normalize(lowScore, minScore, maxScore, lowerBoundScore);
+        float lowScore = 0.2f;
+        float normalizedLowScore = MinMaxScoreNormalizationTechnique.LowerBound.Mode.APPLY.normalize(
+            lowScore,
+            minScore,
+            maxScore,
+            lowerBoundScore
+        );
+        // we expect score as 0.2 - 0.1 / 0.8 - 0.1 = 0.1 / 0.7 = 0.1
         assertEquals(0.143f, normalizedLowScore, DELTA_FOR_SCORE_ASSERTION);
     }
 
-    public void testMode_normalize_clip() {
+    public void testLowerBounds_whenModeIsClip_thenSuccessful() {
         float score = 0.5f;
         float minScore = 0.2f;
         float maxScore = 0.8f;
         float lowerBoundScore = 0.3f;
 
-        float normalizedScore = MinMaxScoreNormalizationTechnique.Mode.CLIP.normalize(score, minScore, maxScore, lowerBoundScore);
+        float normalizedScore = MinMaxScoreNormalizationTechnique.LowerBound.Mode.CLIP.normalize(
+            score,
+            minScore,
+            maxScore,
+            lowerBoundScore
+        );
         assertEquals(0.4f, normalizedScore, DELTA_FOR_SCORE_ASSERTION);
 
         // Test when score is below min score
         float lowScore = 0.1f;
-        float normalizedLowScore = MinMaxScoreNormalizationTechnique.Mode.CLIP.normalize(lowScore, minScore, maxScore, lowerBoundScore);
-        assertEquals(0.6f, normalizedLowScore, DELTA_FOR_SCORE_ASSERTION);
+        float normalizedLowScore = MinMaxScoreNormalizationTechnique.LowerBound.Mode.CLIP.normalize(
+            lowScore,
+            minScore,
+            maxScore,
+            lowerBoundScore
+        );
+        assertEquals(0.0f, normalizedLowScore, DELTA_FOR_SCORE_ASSERTION);
     }
 
-    public void testMode_normalize_ignore() {
+    public void testLowerBounds_whenModeIsIgnore_thenSuccessful() {
         float score = 0.5f;
         float minScore = 0.2f;
         float maxScore = 0.8f;
         float lowerBoundScore = 0.3f;
 
-        float normalizedScore = MinMaxScoreNormalizationTechnique.Mode.IGNORE.normalize(score, minScore, maxScore, lowerBoundScore);
+        float normalizedScore = MinMaxScoreNormalizationTechnique.LowerBound.Mode.IGNORE.normalize(
+            score,
+            minScore,
+            maxScore,
+            lowerBoundScore
+        );
         assertEquals(0.5f, normalizedScore, DELTA_FOR_SCORE_ASSERTION);
 
         // Test when normalized score would be 0
         float lowScore = 0.2f;
-        float normalizedLowScore = MinMaxScoreNormalizationTechnique.Mode.IGNORE.normalize(lowScore, minScore, maxScore, lowerBoundScore);
+        float normalizedLowScore = MinMaxScoreNormalizationTechnique.LowerBound.Mode.IGNORE.normalize(
+            lowScore,
+            minScore,
+            maxScore,
+            lowerBoundScore
+        );
         assertEquals(MIN_SCORE, normalizedLowScore, DELTA_FOR_SCORE_ASSERTION);
     }
 
-    public void testMode_defaultValue() {
-        assertEquals(MinMaxScoreNormalizationTechnique.Mode.APPLY, MinMaxScoreNormalizationTechnique.Mode.DEFAULT);
+    public void testLowerBoundsMode_whenDefaultValue_thenSuccessful() {
+        assertEquals(MinMaxScoreNormalizationTechnique.LowerBound.Mode.APPLY, MinMaxScoreNormalizationTechnique.LowerBound.Mode.DEFAULT);
     }
 
-    public void testLowerBoundsExceedsMaxSubQueries() {
+    public void testLowerBounds_whenExceedsMaxSubQueries_thenFail() {
         List<Map<String, Object>> lowerBounds = new ArrayList<>();
 
         for (int i = 0; i <= 100; i++) {
@@ -375,7 +414,7 @@ public class MinMaxScoreNormalizationTechniqueTests extends OpenSearchQueryTestC
 
         IllegalArgumentException exception = expectThrows(
             IllegalArgumentException.class,
-            () -> new MinMaxScoreNormalizationTechnique(parameters)
+            () -> new MinMaxScoreNormalizationTechnique(parameters, new ScoreNormalizationUtil())
         );
 
         assertEquals(
@@ -387,6 +426,203 @@ public class MinMaxScoreNormalizationTechniqueTests extends OpenSearchQueryTestC
             ),
             exception.getMessage()
         );
+    }
+
+    public void testDescribe_whenLowerBoundsArePresent_thenSuccessful() {
+        Map<String, Object> parameters = new HashMap<>();
+        List<Map<String, Object>> lowerBounds = Arrays.asList(
+            Map.of("mode", "apply", "min_score", 0.2),
+
+            Map.of("mode", "clip", "min_score", 0.1)
+        );
+        parameters.put("lower_bounds", lowerBounds);
+        MinMaxScoreNormalizationTechnique techniqueWithBounds = new MinMaxScoreNormalizationTechnique(
+            parameters,
+            new ScoreNormalizationUtil()
+        );
+        assertEquals("min_max, lower bounds [(apply, 0.2), (clip, 0.1)]", techniqueWithBounds.describe());
+
+        // Test case 2: without lower bounds
+        Map<String, Object> emptyParameters = new HashMap<>();
+        MinMaxScoreNormalizationTechnique techniqueWithoutBounds = new MinMaxScoreNormalizationTechnique(
+            emptyParameters,
+            new ScoreNormalizationUtil()
+        );
+        assertEquals("min_max", techniqueWithoutBounds.describe());
+
+        Map<String, Object> parametersMissingMode = new HashMap<>();
+        List<Map<String, Object>> lowerBoundsMissingMode = Arrays.asList(
+            Map.of("min_score", 0.2),
+            Map.of("mode", "clip", "min_score", 0.1)
+        );
+        parametersMissingMode.put("lower_bounds", lowerBoundsMissingMode);
+        MinMaxScoreNormalizationTechnique techniqueMissingMode = new MinMaxScoreNormalizationTechnique(
+            parametersMissingMode,
+            new ScoreNormalizationUtil()
+        );
+        assertEquals("min_max, lower bounds [(apply, 0.2), (clip, 0.1)]", techniqueMissingMode.describe());
+
+        Map<String, Object> parametersMissingScore = new HashMap<>();
+        List<Map<String, Object>> lowerBoundsMissingScore = Arrays.asList(
+            Map.of("mode", "apply"),
+            Map.of("mode", "clip", "min_score", 0.1)
+        );
+        parametersMissingScore.put("lower_bounds", lowerBoundsMissingScore);
+        MinMaxScoreNormalizationTechnique techniqueMissingScore = new MinMaxScoreNormalizationTechnique(
+            parametersMissingScore,
+            new ScoreNormalizationUtil()
+        );
+        assertEquals("min_max, lower bounds [(apply, 0.0), (clip, 0.1)]", techniqueMissingScore.describe());
+    }
+
+    public void testLowerBounds_whenInvalidInput_thenFail() {
+        // Test case 1: Invalid mode value
+        Map<String, Object> parametersInvalidMode = new HashMap<>();
+        List<Map<String, Object>> lowerBoundsInvalidMode = Arrays.asList(
+            Map.of("mode", "invalid_mode", "min_score", 0.2),
+            Map.of("mode", "clip", "min_score", 0.1)
+        );
+        parametersInvalidMode.put("lower_bounds", lowerBoundsInvalidMode);
+        IllegalArgumentException invalidModeException = expectThrows(
+            IllegalArgumentException.class,
+            () -> new MinMaxScoreNormalizationTechnique(parametersInvalidMode, new ScoreNormalizationUtil())
+        );
+        assertEquals("invalid mode: invalid_mode, valid values are: apply, clip, ignore", invalidModeException.getMessage());
+
+        // Test case 4: Invalid min_score type
+        Map<String, Object> parametersInvalidScore = new HashMap<>();
+        List<Map<String, Object>> lowerBoundsInvalidScore = Arrays.asList(
+            Map.of("mode", "apply", "min_score", "not_a_number"),
+            Map.of("mode", "clip", "min_score", 0.1)
+        );
+        parametersInvalidScore.put("lower_bounds", lowerBoundsInvalidScore);
+        IllegalArgumentException invalidScoreException = expectThrows(
+            IllegalArgumentException.class,
+            () -> new MinMaxScoreNormalizationTechnique(parametersInvalidScore, new ScoreNormalizationUtil())
+        );
+        assertEquals("invalid format for min_score: must be a valid float value", invalidScoreException.getMessage());
+    }
+
+    public void testLowerBoundsValidation_whenLowerBoundsAndSubQueriesCountMismatch_thenFail() {
+        Map<String, Object> parameters = new HashMap<>();
+        List<Map<String, Object>> lowerBounds = Arrays.asList(Map.of("mode", "clip", "min_score", 0.1));
+        parameters.put(PARAM_NAME_LOWER_BOUNDS, lowerBounds);
+
+        List<CompoundTopDocs> compoundTopDocs = List.of(
+            new CompoundTopDocs(
+                new TotalHits(2, TotalHits.Relation.EQUAL_TO),
+                List.of(
+                    new TopDocs(
+                        new TotalHits(2, TotalHits.Relation.EQUAL_TO),
+                        new ScoreDoc[] { new ScoreDoc(2, 0.5f), new ScoreDoc(4, 0.2f) }
+                    ),
+                    new TopDocs(new TotalHits(1, TotalHits.Relation.EQUAL_TO), new ScoreDoc[] { new ScoreDoc(3, 0.1f) })
+                ),
+                false,
+                SEARCH_SHARD
+            )
+        );
+        ScoreNormalizationTechnique minMaxTechnique = new MinMaxScoreNormalizationTechnique(parameters, new ScoreNormalizationUtil());
+        NormalizeScoresDTO normalizeScoresDTO = NormalizeScoresDTO.builder()
+            .queryTopDocs(compoundTopDocs)
+            .normalizationTechnique(minMaxTechnique)
+            .build();
+
+        IllegalArgumentException exception = expectThrows(
+            IllegalArgumentException.class,
+            () -> minMaxTechnique.normalize(normalizeScoresDTO)
+        );
+
+        assertEquals(
+            "expected lower bounds array to contain 2 elements matching the number of sub-queries, but found a mismatch",
+            exception.getMessage()
+        );
+    }
+
+    public void testLowerBoundsValidation_whenTopDocsIsEmpty_thenSuccessful() {
+        Map<String, Object> parameters = new HashMap<>();
+        List<Map<String, Object>> lowerBounds = Arrays.asList(
+            Map.of("mode", "clip", "min_score", 0.1),
+            Map.of("mode", "apply", "min_score", 0.0)
+        );
+        parameters.put(PARAM_NAME_LOWER_BOUNDS, lowerBounds);
+
+        List<CompoundTopDocs> compoundTopDocs = List.of(
+            new CompoundTopDocs(new TotalHits(0, TotalHits.Relation.EQUAL_TO), List.of(), false, SEARCH_SHARD),
+            new CompoundTopDocs(
+                new TotalHits(2, TotalHits.Relation.EQUAL_TO),
+                List.of(
+                    new TopDocs(
+                        new TotalHits(2, TotalHits.Relation.EQUAL_TO),
+                        new ScoreDoc[] { new ScoreDoc(2, 0.5f), new ScoreDoc(4, 0.2f) }
+                    ),
+                    new TopDocs(new TotalHits(1, TotalHits.Relation.EQUAL_TO), new ScoreDoc[] { new ScoreDoc(3, 0.1f) })
+                ),
+                false,
+                SEARCH_SHARD
+            )
+        );
+        ScoreNormalizationTechnique minMaxTechnique = new MinMaxScoreNormalizationTechnique(parameters, new ScoreNormalizationUtil());
+        NormalizeScoresDTO normalizeScoresDTO = NormalizeScoresDTO.builder()
+            .queryTopDocs(compoundTopDocs)
+            .normalizationTechnique(minMaxTechnique)
+            .build();
+
+        minMaxTechnique.normalize(normalizeScoresDTO);
+
+        CompoundTopDocs expectedCompoundDocsZero = new CompoundTopDocs(
+            new TotalHits(0, TotalHits.Relation.EQUAL_TO),
+            List.of(),
+            false,
+            SEARCH_SHARD
+        );
+        CompoundTopDocs expectedCompoundDocsOne = new CompoundTopDocs(
+            new TotalHits(2, TotalHits.Relation.EQUAL_TO),
+            List.of(
+                new TopDocs(
+                    new TotalHits(2, TotalHits.Relation.EQUAL_TO),
+                    new ScoreDoc[] { new ScoreDoc(2, 1.0f), new ScoreDoc(4, 0.25f) }
+                ),
+                new TopDocs(new TotalHits(1, TotalHits.Relation.EQUAL_TO), new ScoreDoc[] { new ScoreDoc(3, 1.0f) })
+            ),
+            false,
+            SEARCH_SHARD
+        );
+        expectedCompoundDocsOne.setScoreDocs(List.of(new ScoreDoc(2, 0.5f), new ScoreDoc(4, 0.2f)));
+        assertNotNull(compoundTopDocs);
+        assertEquals(2, compoundTopDocs.size());
+        CompoundTopDocs compoundTopDocsZero = compoundTopDocs.get(0);
+        assertEquals(expectedCompoundDocsZero, compoundTopDocsZero);
+        CompoundTopDocs compoundTopDocsOne = compoundTopDocs.get(1);
+        assertEquals(expectedCompoundDocsOne, compoundTopDocsOne);
+    }
+
+    public void testInvalidParameters() {
+        Map<String, Object> parameters = new HashMap<>();
+        List<Map<String, Object>> lowerBoundsList = List.of(
+            Map.of("min_score", 0.1, "mode", "clip"),
+            Map.of("mode", "ignore", "invalid_param", "value") // Adding an invalid nested parameter
+        );
+        parameters.put("lower_bounds", lowerBoundsList);
+
+        try {
+            new MinMaxScoreNormalizationTechnique(parameters, new ScoreNormalizationUtil());
+            fail("Expected IllegalArgumentException was not thrown");
+        } catch (IllegalArgumentException e) {
+            assertEquals("unrecognized parameters in normalization technique", e.getMessage());
+        }
+    }
+
+    public void testUnsupportedTopLevelParameter() {
+        Map<String, Object> parameters = new HashMap<>();
+        parameters.put("invalid_top_level_param", "value"); // Adding an invalid top-level parameter
+
+        try {
+            new MinMaxScoreNormalizationTechnique(parameters, new ScoreNormalizationUtil());
+            fail("Expected IllegalArgumentException was not thrown");
+        } catch (IllegalArgumentException e) {
+            assertEquals("unrecognized parameters in normalization technique", e.getMessage());
+        }
     }
 
     private void assertCompoundTopDocs(TopDocs expected, TopDocs actual) {